Changed program counter to word. Removed and re-added .h file since originally cerating it as a .c file and changing the extension created compilation issues.

+++ DLX-Sim/DLX-Sim/register_file_handler.h
@@ -9,7 +9,7 @@ typedef struct
   WORD gpr[32];
 	// 32 Floating Point Registers
   WORD fpr[32];
-  DWORD program_counter;
+  WORD program_counter;
 }Registers;
 
 extern Registers* registers;
